- Se añaden operaciones de busqueda, creación y actualización de categorias - Se añade al ProductDto categories

+++ MarketNet/src/Infraestructure/Repositories/IRepository.cs
@@ -2,7 +2,7 @@
 {
     public interface IRepository<TEntity> where TEntity : class
     {
-        Task<TEntity?> GetByIdAsync(int id);
+        Task<TEntity?> GetByIdAsync(long id);
         Task<IEnumerable<TEntity>> GetAllAsync();
         Task<long> AddAsync(TEntity entity);
         Task UpdateAsync(TEntity entity);
