+++ MarketNet/src/Infraestructure/Repositories/Impl/GenericRepository.cs
@@ -17,7 +17,7 @@ namespace MarketNet.src.Infraestructure.Repositories.Impl
             _dbSet = _context.Set<TEntity>();
         }
 
-        public async Task<TEntity?> GetByIdAsync(int id)
+        public async Task<TEntity?> GetByIdAsync(long id)
         {
             return await _dbSet.FindAsync(id);
         }
